+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - `IInventory<T>` and `Inventory<T>`
   - `CanAdd(T item)`
-- 
+- `IStackInventory<T>` and `StackInventory<T>`
+  - `CanAdd(T item)`
 
 ## Known Issues
 * **The Current Architecture is not stable for the final version yet**